reorganize jax backend APIs and docs
