+++ brainpy/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = "1.1.1"
+__version__ = "1.1.2"
 
 try:
   import jax
