+++ brainpy/math/jax/ops.py
@@ -3,8 +3,6 @@
 import jax.numpy as jnp
 import numpy as np
 from jax.tree_util import tree_map
-from jax import ops as jops
-from jax import jit
 
 from brainpy.math.jax.jaxarray import JaxArray
 from brainpy.math.numpy import ops
@@ -74,50 +72,12 @@ __all__ = [
   'float64', 'complex_', 'complex64', 'complex128', 'set_int_', 'set_float_', 'set_complex_',
 
   # others
-  'take_along_axis', 'clip_by_norm', 'segment_sum',
-  'segment_prod', 'segment_max', 'segment_min',
+  'take_along_axis', 'clip_by_norm',
 ]
 
 
 # others
 # ------
-
-jit_seg_sum = jit(jops.segment_sum, static_argnums=(2, 3, 4, 5))
-jit_seg_prod = jit(jops.segment_prod, static_argnums=(2, 3, 4, 5))
-jit_seg_max = jit(jops.segment_max, static_argnums=(2, 3, 4, 5))
-jit_seg_min = jit(jops.segment_min, static_argnums=(2, 3, 4, 5))
-
-
-def segment_sum(data, segment_ids, num_segments: int, indices_are_sorted: bool = False,
-                unique_indices: bool = False, bucket_size: int = None):
-  """Computes the sum within segments of an array."""
-  data = data.value if isinstance(data, JaxArray) else data
-  segment_ids = segment_ids.value if isinstance(segment_ids, JaxArray) else segment_ids
-  return jit_seg_sum(data, segment_ids, num_segments, indices_are_sorted, unique_indices, bucket_size)
-
-
-def segment_prod(data, segment_ids, num_segments: int, indices_are_sorted: bool = False,
-                unique_indices: bool = False, bucket_size: int = None):
-  """Computes the product within segments of an array."""
-  data = data.value if isinstance(data, JaxArray) else data
-  segment_ids = segment_ids.value if isinstance(segment_ids, JaxArray) else segment_ids
-  return jit_seg_prod(data, segment_ids, num_segments, indices_are_sorted, unique_indices, bucket_size)
-
-
-def segment_max(data, segment_ids, num_segments: int, indices_are_sorted: bool = False,
-                unique_indices: bool = False, bucket_size: int = None):
-  """Computes the product within segments of an array."""
-  data = data.value if isinstance(data, JaxArray) else data
-  segment_ids = segment_ids.value if isinstance(segment_ids, JaxArray) else segment_ids
-  return jit_seg_max(data, segment_ids, num_segments, indices_are_sorted, unique_indices, bucket_size)
-
-
-def segment_min(data, segment_ids, num_segments: int, indices_are_sorted: bool = False,
-                unique_indices: bool = False, bucket_size: int = None):
-  """Computes the product within segments of an array."""
-  data = data.value if isinstance(data, JaxArray) else data
-  segment_ids = segment_ids.value if isinstance(segment_ids, JaxArray) else segment_ids
-  return jit_seg_min(data, segment_ids, num_segments, indices_are_sorted, unique_indices, bucket_size)
 
 
 @copy_doc(np.take_along_axis)
@@ -131,8 +91,6 @@ def take_along_axis(a, indices, axis):
 def clip_by_norm(t, clip_norm, axis=None):
   f = lambda l: l * clip_norm / maximum(sqrt(sum(l * l, axis=axis, keepdims=True)), clip_norm)
   return tree_map(f, t)
-
-
 
 
 # math funcs
