+++ docs/apis/math/special_jax.rst
@@ -5,9 +5,12 @@ JAX Special Supports
 .. toctree::
     :maxdepth: 2
 
+    jax_math/parallels
+    jax_math/operators
     jax_math/controls
     jax_math/autograd
     jax_math/activations
+    jax_math/settings
     jax_math/losses
     jax_math/optimizers
